web signup issues fixed

+++ cms-project-web/src/globalClasses/WorkoutDetails.js
@@ -1,12 +1,21 @@
 import React from "react";
 import { useWorkoutsContext } from "../Hook/useWorkoutsContext";
 import formatDistanceToNow from "date-fns/formatDistanceToNow";
+import { useAuthContext } from "../Hook/useAuthContext";
 
 const WorkoutDetails = ({ workout }) => {
+  const { user } = useAuthContext();
   const { dispatch } = useWorkoutsContext();
   const handleClick = async () => {
+    if (!user) {
+      return;
+    }
+
     const response = await fetch("api/workouts/" + workout._id, {
       method: "DELETE",
+      headers: {
+        Authorization: `Bearer ${user.token}`,
+      },
     });
 
     const json = await response.json();
